Menu component + redesigning the HOC
