+++ src/Routes/About/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import withRouteComponent from '../withRouteComponent';
+import withRouteComponent from '../utils/withRouteComponent';
 
 class About extends React.Component {
   render() {
